+++ packages/theming/default-theme/src/defaultTheme.ts
@@ -4,7 +4,8 @@ import { globalTokens } from '@fluentui-react-native/theme-tokens';
 import type { Theme, Typography, Spacing, FontWeightValue, FontSize, FontSizes, Variants } from '@fluentui-react-native/theme-types';
 
 import { createShadowAliasTokens } from './createAliasTokens';
-import { getStockWebPalette, getStockWebDarkPalette, getStockWebHCPalette } from './defaultColors';
+import { getStockWebHCPalette } from './defaultColors';
+import { getStockWebPalette, getStockWebDarkPalette } from './stockWebPalette';
 
 function _defaultTypography(): Typography {
   const defaultsDict = {
